Rename ui -> master_ui.
