+++ tunnels/src/clock.rs
@@ -1,5 +1,5 @@
+use crate::master_ui::EmitStateChange as EmitShowStateChange;
 use crate::numbers::UnipolarFloat;
-use crate::ui::EmitStateChange as EmitShowStateChange;
 use serde::{Deserialize, Serialize};
 use std::time::{Duration, Instant};
